Rename the new files since they clash with the memory.hpp in utility
