+++ src/CMakeLists.txt
@@ -13,7 +13,7 @@ add_executable(
     compiler.cpp
     program.cpp
     runtime.cpp
-    memory.cpp
+    allocator.cpp
     object.cpp
     functions.cpp
     operators.cpp
